Related_Products/requirements.txt
flask
psycopg2-binary
pandas
faiss-cpu
sentence-transformers
